requirements.txt
# Core ML and AI dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
accelerate>=0.20.0

# ML and data processing
scikit-learn>=1.3.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# Web interface
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
